+++ DesktopApplicationTemplate.UI/Factories/IServiceFactory.cs
@@ -1,10 +1,11 @@
 using DesktopApplicationTemplate.UI.ViewModels;
+using DesktopApplicationTemplate.Models;
 
 namespace DesktopApplicationTemplate.UI.Factories
 {
     public interface IServiceFactory
     {
-        string ServiceType { get; }
+        ServiceType ServiceType { get; }
         ServiceListModel Create(object options);
     }
 }